Implement endpoint that loads the commentary detail view for a specific version of a commentary using a given revision timestamp.
Refactor commentary component to reload the detail view when a specific version of a commentary is selected instead of asynchronously loading commentary data inline. Note: Reloading the entire view allows dynamic components within the selected version content to be rendered correctly, which would not be possible when requesting the version data asynchronously.

Set the active version based on the selected version timestamp.

Display dates of commentary versions in correct format based on the active locale. Note: Dates in English are displayed as "MM.DD.YYYY". Dates in German, French and Italian are displayed as "DD.MM.YYYY".