Display the footnote list at the bottom of a commentary on print.
Add a new store file making use of the Vue 3 reactivity API (see https://vuejs.org/guide/scaling-up/state-management.html#simple-state-management-with-reactivity-api) to store an array of footnotes, which gets populated by each Footnote component inside the commentary content when mounted. Use the list to fill in a list located at the bottom of the Commentary component, which is hidden on screen but shown when printed.

Add translation lines  for the footnotes section header.