Disable git automation in local environment.
Remove git package installation in PHP container for local development. Disable git in .env.example file for Statamic. In the context of this application setup, git automation should only be used to track changes to the data folder (`sources/webapp/data/`) on the remote server (staging or production), where they could be automatically committed and pushed. The changes should be retrieved manually by developers, to limit the risks of accidentally pushing local changes to the data to the dedicated remote repository.